added latest development version of SDLTris
too many changes to note

+++ source/font.h
@@ -4,10 +4,9 @@
 
 class Font {
     public:
-        virtual void render(std::string words, int x, int y,
-        bool center, int red, int blue, int green,
-        int wordwrap, bool sine, double pos,
-        double multiplyin, double multiplyout, int scale){};
+        virtual void render(std::string words, int x, int y,bool center, 
+        int red, int blue, int green, int wordwrap, bool sine, double pos,
+        double multiplyin, double multiplyout, double scale){};
         
         virtual void render(int x, int y,std::string words,
         bool center, int red, int blue, int green,
